+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 - Disabled orderDetailDisplayed position, because it dont have function
 
+### Improved
+- Added compatibility with module onepagecheckoutps (v4.2.3)
+
 ## [1.2.13]
 ### Fixed
 - Fixed block loading crash when getting an error when trying to register a shipment from Itella in the block order view